backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai>=0.8.5
av>=12.0.0
faster-whisper>=1.0.0
sounddevice==0.4.6
soundfile==0.12.1
numpy>=1.26.0
pymongo==4.6.0
pyjwt==2.8.0
bcrypt==4.1.2

